Fix large type integer/float conversion and add multi-block static_assert
cfloat changes:
- Fix setfraction() to work for fbits >= 64 (was a no-op)
- Fix convert_signed_integer to place fraction bits at TOP of fraction
  field for large types (fbits >= 64 - es)
- Fix round() to not zero raw when shift >= 64

areal changes:
- Add static_assert preventing uint64_t blocks for multi-block types
  (carry propagation requires 32-bit or smaller blocks)
- Fix double-to-areal shift overflow for large types (shiftAmount > 12)
- Fix fraction bit placement for large areals (> 64 bits) to place
  bits at TOP of fraction field instead of bottom

muller.cpp:
- Update areal<128,15> and areal<256,19> to use uint32_t blocks

Root cause: Integer values like 111 were being stored as 64 (power of 2
truncation) because fraction bits were placed at wrong positions.

+++ applications/precision/ubit/muller.cpp
@@ -129,8 +129,8 @@ try {
 	std::cout << "\n=== areal (with ubit uncertainty tracking) ===";
 	test_muller<areal<32, 8, uint32_t>>("areal<32,8>", max_iterations);
 	test_muller<areal<64, 11, uint64_t>>("areal<64,11>", max_iterations);
-	test_muller<areal<128, 15, uint64_t>>("areal<128,15>", max_iterations);
-	test_muller<areal<256, 19, uint64_t>>("areal<256,19>", max_iterations);
+	test_muller<areal<128, 15, uint32_t>>("areal<128,15>", max_iterations);
+	test_muller<areal<256, 19, uint32_t>>("areal<256,19>", max_iterations);
 
 	std::cout << "\n" << std::string(80, '=') << '\n';
 	std::cout << "Key insight:\n";
